replace colour references with color
