+++ packages/perspective-viewer-d3fc/src/js/charts/line.js
@@ -9,10 +9,10 @@
 import * as fc from "d3fc";
 import * as crossAxis from "../axis/crossAxis";
 import * as mainAxis from "../axis/mainAxis";
-import {seriesColours} from "../series/seriesColours";
+import {seriesColors} from "../series/seriesColors";
 import {lineSeries} from "../series/lineSeries";
 import {splitData} from "../data/splitData";
-import {colourLegend} from "../legend/legend";
+import {colorLegend} from "../legend/legend";
 import {filterData} from "../legend/filter";
 import {withGridLines} from "../gridlines/gridlines";
 
@@ -22,13 +22,13 @@ import nearbyTip from "../tooltip/nearbyTip";
 
 function lineChart(container, settings) {
     const data = splitData(settings, filterData(settings));
-    const colour = seriesColours(settings);
+    const color = seriesColors(settings);
 
-    const legend = colourLegend()
+    const legend = colorLegend()
         .settings(settings)
-        .scale(colour);
+        .scale(color);
 
-    const series = fc.seriesSvgRepeat().series(lineSeries(settings, colour).orient("vertical"));
+    const series = fc.seriesSvgRepeat().series(lineSeries(settings, color).orient("vertical"));
 
     const paddingStrategy = hardLimitZeroPadding()
         .pad([0.1, 0.1])
@@ -68,7 +68,7 @@ function lineChart(container, settings) {
         .settings(settings)
         .xScale(xScale)
         .yScale(yScale)
-        .colour(colour)
+        .color(color)
         .data(data);
     container.call(toolTip);
