+++ packages/perspective-viewer-d3fc/src/js/charts/y-scatter.js
@@ -9,7 +9,7 @@
 import * as fc from "d3fc";
 import * as crossAxis from "../axis/crossAxis";
 import * as mainAxis from "../axis/mainAxis";
-import {seriesColours} from "../series/seriesColours";
+import {seriesColors} from "../series/seriesColors";
 import {categoryPointSeries, symbolType} from "../series/categoryPointSeries";
 import {groupData} from "../data/groupData";
 import {symbolLegend} from "../legend/legend";
@@ -21,17 +21,17 @@ import zoomableChart from "../zoom/zoomableChart";
 function yScatter(container, settings) {
     const data = groupData(settings, filterData(settings));
     const symbols = symbolType(settings);
-    const colour = seriesColours(settings);
+    const color = seriesColors(settings);
 
     const legend = symbolLegend()
         .settings(settings)
         .scale(symbols)
-        .colour(colour);
+        .color(color);
 
     const series = fc
         .seriesSvgMulti()
         .mapping((data, index) => data[index])
-        .series(data.map(series => categoryPointSeries(settings, series.key, colour, symbols)));
+        .series(data.map(series => categoryPointSeries(settings, series.key, color, symbols)));
 
     const paddingStrategy = hardLimitZeroPadding()
         .pad([0.05, 0.05])
